requirements.txt
pip install --upgrade python-telegram-bot
openai==1.*
gspread==5.*
python-dotenv
pandas